remove TuiWindow children heirarchy
Remove the parent-child relationship between TuiWindow and TuiAtlas, TuiPalette, TuiPanel, and TuiTexture by removing TuiWindow as a parameter to these structs.
Remove counters of each of these objects from TuiWindow struct.
Add static counters to c files of each object instead.
Add tuiObjectGetCount functions to get the amount of these objects that currently exist.
remove dangling error object checks from tuiWindowDestroy

+++ src/cursor.c
@@ -6,7 +6,7 @@
 #include <TUIC/system.h>
 #include <TUIC/boolean.h>
 
-static size_t sCursorCount;
+static int sCursorCount;
 
 TuiCursor tuiCursorCreate(TuiImage image, int xhot, int yhot)
 {
@@ -56,5 +56,5 @@ void tuiCursorDestroy(TuiCursor cursor)
 
 int tuiGetCursorCount()
 {
-	return (int)sCursorCount;
+	return sCursorCount;
 }
